calendar fixes

+++ routers/staff/members/data_structures.py
@@ -8,9 +8,9 @@ class Member(BaseModel):
     email: str
     sites: Optional[list[str]] = []
     admin: Optional[bool] = False
-    al_entitlement: int = 25
-    al_used: int = 0
-    al_remaining: int = 25
+    al_entitlement: float = 25.0
+    al_used: float = 0.0
+    al_remaining: float = 25.0
     sickness_used: int = 0
     username: str
